old-tests/submission/W3C/domhtml/HTMLObjectElement04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OBJECT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-object-border" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<object border="0" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<script>
var domtest = async_test("The border IDL attribute of the object element must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vborder;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("object");
      assert_equals(nodeList.length,1);
      testNode = nodeList.item(0);
      vborder = testNode.border;

      assert_equals(vborder,"0");
       
    domtest.done();
}));

</script>
<p>The border IDL attribute of the object element must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
